test: 기존 게임 로직 테스트 코드 구조 개선
