+++ BE/test/integration/quiz-set/quiz.integration.spec.ts
@@ -3,19 +3,19 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { DataSource, QueryRunner } from 'typeorm';
 import { BadRequestException } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { QuizSetService } from '../../src/quiz-set/service/quiz-set.service';
-import { QuizSetModel } from '../../src/quiz-set/entities/quiz-set.entity';
-import { QuizModel } from '../../src/quiz-set/entities/quiz.entity';
-import { QuizChoiceModel } from '../../src/quiz-set/entities/quiz-choice.entity';
-import { UserModel } from '../../src/user/entities/user.entity';
-import { UserQuizArchiveModel } from '../../src/user/entities/user-quiz-archive.entity';
-import { CreateQuizSetDto } from '../../src/quiz-set/dto/create-quiz.dto';
-import { QuizSetCreateService } from '../../src/quiz-set/service/quiz-set-create.service';
-import { QuizSetReadService } from '../../src/quiz-set/service/quiz-set-read.service';
-import { QuizSetUpdateService } from '../../src/quiz-set/service/quiz-set-update.service';
-import { QuizSetDeleteService } from '../../src/quiz-set/service/quiz-set-delete.service';
-import { UserService } from '../../src/user/user.service';
-import { AuthModule } from '../../src/auth/auth.module';
+import { QuizSetService } from '../../../src/quiz-set/service/quiz-set.service';
+import { QuizSetModel } from '../../../src/quiz-set/entities/quiz-set.entity';
+import { QuizModel } from '../../../src/quiz-set/entities/quiz.entity';
+import { QuizChoiceModel } from '../../../src/quiz-set/entities/quiz-choice.entity';
+import { UserModel } from '../../../src/user/entities/user.entity';
+import { UserQuizArchiveModel } from '../../../src/user/entities/user-quiz-archive.entity';
+import { CreateQuizSetDto } from '../../../src/quiz-set/dto/create-quiz.dto';
+import { QuizSetCreateService } from '../../../src/quiz-set/service/quiz-set-create.service';
+import { QuizSetReadService } from '../../../src/quiz-set/service/quiz-set-read.service';
+import { QuizSetUpdateService } from '../../../src/quiz-set/service/quiz-set-update.service';
+import { QuizSetDeleteService } from '../../../src/quiz-set/service/quiz-set-delete.service';
+import { UserService } from '../../../src/user/user.service';
+import { AuthModule } from '../../../src/auth/auth.module';
 
 describe('QuizService', () => {
   let quizService: QuizSetService;
